Fix standalone-header CI breaks from pruned transitive includes
Two header self-containment regressions surfaced by CI (clang-tidy on
changed files + MSVC build):

- csv_casting.hpp used CastParameters by value but only got the forward
  declaration from cast_operators.hpp; include default_casts.hpp.
- serialization_data.hpp holds stack<const_reference<CompressionInfo>>,
  which MSVC eagerly instantiates and requires CompressionInfo complete
  (clang instantiates lazily, so it passed locally). CompressionInfo lived
  in the heavy compression_function.hpp hub (pulls client_context.hpp), so
  extract it into a light compression_info.hpp (only needs BlockManager)
  and include that instead of forward-declaring.

Verified: all 115 branch-changed headers now compile standalone.
