+++ src/include/duckdb/execution/operator/csv_scanner/csv_casting.hpp
@@ -12,6 +12,7 @@
 #include "duckdb/common/vector_operations/unary_executor.hpp"
 #include "duckdb/common/operator/cast_operators.hpp"
 #include "duckdb/common/operator/decimal_cast_operators.hpp"
+#include "duckdb/function/cast/default_casts.hpp"
 
 namespace duckdb {
 struct CastParameters;
